requirements.txt
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cohere==4.48
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.109.2
fastavro==1.9.4
filelock==3.13.1
fiona==1.9.5
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
geopandas==0.14.3
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kiwisolver==1.4.5
lida==0.0.14
llmx==0.0.21a0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-venn==0.11.10
mdurl==0.1.2
mizani==0.11.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
openai==1.12.0
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
patsy==0.5.6
pillow==10.2.0
plotly==5.19.0
plotnine==0.13.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.36.3
statsmodels==0.14.1
streamlit==1.31.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.0
tornado==6.4
tqdm==4.66.2
transformers==4.37.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.27.1
validators==0.22.0
watchdog==4.0.0
wordcloud==1.9.3
yarl==1.9.4
zipp==3.17.0
